+++ src/main/java/miyucomics/hexcassettes/HexcassettesMain.kt
@@ -1,7 +1,6 @@
 package miyucomics.hexcassettes
 
 import at.petrak.hexcasting.api.HexAPI
-import miyucomics.hexcassettes.HexcassettesUtils.id
 import miyucomics.hexcassettes.inits.HexcassettesAdvancements
 import miyucomics.hexcassettes.inits.HexcassettesNetworking
 import miyucomics.hexcassettes.inits.HexcassettesPatterns
@@ -17,6 +16,7 @@ import net.minecraft.registry.Registry
 import net.minecraft.registry.RegistryKey
 import net.minecraft.server.network.ServerPlayerEntity
 import net.minecraft.sound.SoundCategory
+import net.minecraft.util.Identifier
 import net.minecraft.util.Rarity
 import net.minecraft.world.World
 
@@ -35,7 +35,7 @@ class HexcassettesMain : ModInitializer {
 	companion object {
 		const val MOD_ID: String = "hexcassettes"
 		const val MAX_CASSETTES: Int = 6
-		const val MAX_LABEL_LENGTH: Int = 32
+		fun id(string: String) = Identifier(MOD_ID, string)
 	}
 }
 
@@ -49,15 +49,17 @@ class CassetteItem : Item(Settings().maxCount(1).rarity(Rarity.UNCOMMON).food(Fo
 			world.playSound(user.x, user.y, user.z, HexcassettesSounds.CASSETTE_INSERT, SoundCategory.MASTER, 5f, 1f, false)
 			return super.finishUsing(stack, world, user)
 		}
+
 		if (user !is ServerPlayerEntity)
 			return super.finishUsing(stack, world, user)
-		val playerState = HexcassettesAPI.getPlayerState(user)
-		if (playerState.ownedCassettes < HexcassettesMain.MAX_CASSETTES) {
+
+		val cassetteData = (user as PlayerEntityMinterface).getCassetteState()
+		if (cassetteData.ownedSlots < HexcassettesMain.MAX_CASSETTES) {
 			HexcassettesAdvancements.TAPE_WORM.trigger(user)
-			HexcassettesAPI.getPlayerState(user).ownedCassettes += 1
-			if (HexcassettesAPI.getPlayerState(user).ownedCassettes == HexcassettesMain.MAX_CASSETTES)
+			cassetteData.ownedSlots += 1
+			if (cassetteData.ownedSlots == HexcassettesMain.MAX_CASSETTES)
 				HexcassettesAdvancements.FULL_ARSENAL.trigger(user)
-			HexcassettesAPI.sendSyncPacket(user)
+			cassetteData.sync(user)
 		}
 		return super.finishUsing(stack, world, user)
 	}