+++ src/main/java/miyucomics/hexcassettes/inits/HexcassettesSounds.kt
@@ -1,6 +1,6 @@
 package miyucomics.hexcassettes.inits
 
-import miyucomics.hexcassettes.HexcassettesUtils
+import miyucomics.hexcassettes.HexcassettesMain
 import net.minecraft.registry.Registries
 import net.minecraft.registry.Registry
 import net.minecraft.sound.SoundEvent
@@ -19,7 +19,7 @@ object HexcassettesSounds {
 	}
 
 	private fun register(name: String): SoundEvent {
-		val id = HexcassettesUtils.id(name)
+		val id = HexcassettesMain.id(name)
 		val event = SoundEvent.of(id)
 		Registry.register(Registries.SOUND_EVENT, id, event)
 		return event