+++ src/main/java/miyucomics/hexcassettes/patterns/OpDequeue.kt
@@ -4,23 +4,23 @@ import at.petrak.hexcasting.api.casting.RenderedSpell
 import at.petrak.hexcasting.api.casting.castables.SpellAction
 import at.petrak.hexcasting.api.casting.eval.CastingEnvironment
 import at.petrak.hexcasting.api.casting.eval.env.PlayerBasedCastEnv
+import at.petrak.hexcasting.api.casting.getPositiveIntUnder
 import at.petrak.hexcasting.api.casting.iota.Iota
 import at.petrak.hexcasting.api.casting.mishaps.MishapBadCaster
-import miyucomics.hexcassettes.HexcassettesAPI
-import miyucomics.hexcassettes.HexcassettesUtils
-import net.minecraft.server.network.ServerPlayerEntity
+import miyucomics.hexcassettes.HexcassettesMain
+import miyucomics.hexcassettes.PlayerEntityMinterface
 
 class OpDequeue : SpellAction {
 	override val argc = 1
 	override fun execute(args: List<Iota>, env: CastingEnvironment): SpellAction.Result {
 		if (env !is PlayerBasedCastEnv)
 			throw MishapBadCaster()
-		return SpellAction.Result(Spell(env.castingEntity as ServerPlayerEntity, HexcassettesUtils.shortenLabel(args[0].display().string)), 0, listOf())
+		return SpellAction.Result(Spell(args.getPositiveIntUnder(0, HexcassettesMain.MAX_CASSETTES, argc)), 0, listOf())
 	}
 
-	private data class Spell(val caster: ServerPlayerEntity, val label: String) : RenderedSpell {
+	private data class Spell(val index: Int) : RenderedSpell {
 		override fun cast(env: CastingEnvironment) {
-			HexcassettesAPI.dequeueByName(caster, label)
+			(env.castingEntity as PlayerEntityMinterface).getCassetteState().queuedHexes[index] = null
 		}
 	}
 }