+++ src/main/java/miyucomics/hexcassettes/patterns/OpFree.kt
@@ -4,17 +4,21 @@ import at.petrak.hexcasting.api.casting.asActionResult
 import at.petrak.hexcasting.api.casting.castables.ConstMediaAction
 import at.petrak.hexcasting.api.casting.eval.CastingEnvironment
 import at.petrak.hexcasting.api.casting.eval.env.PlayerBasedCastEnv
+import at.petrak.hexcasting.api.casting.iota.DoubleIota
 import at.petrak.hexcasting.api.casting.iota.Iota
 import at.petrak.hexcasting.api.casting.mishaps.MishapBadCaster
-import miyucomics.hexcassettes.HexcassettesAPI
-import net.minecraft.server.network.ServerPlayerEntity
+import miyucomics.hexcassettes.PlayerEntityMinterface
 
 class OpFree : ConstMediaAction {
 	override val argc = 0
 	override fun execute(args: List<Iota>, env: CastingEnvironment): List<Iota> {
 		if (env !is PlayerBasedCastEnv)
 			throw MishapBadCaster()
-		val playerState = HexcassettesAPI.getPlayerState((env.castingEntity as ServerPlayerEntity))
-		return (playerState.ownedCassettes - playerState.queuedHexes.size).asActionResult
+		val freeSlots = mutableListOf<DoubleIota>()
+		(env.castingEntity as PlayerEntityMinterface).getCassetteState().queuedHexes.forEachIndexed { index, hex ->
+			if (hex == null)
+				freeSlots.add(DoubleIota(index.toDouble()))
+		}
+		return freeSlots.asActionResult
 	}
 }